Scrape all open "add wiki" issues from IWB GitHub
Use the GitHub API to get a list of all open "add wiki" issues for IWB,
then parse those issues to get the URLs of the proposed origin and
destination wikis.

Issues can have multiple origins, but can only have a single
destination. If an issue proposes multiple destinations, it will fail
to parse.

The scraped data is exported to a CSV file.